refactor: move session types to their own module
Address review: ServiceId, SessionId, and SessionTarget move from
network_types::types into a dedicated network_types::session module
(still re-exported at network_types for ergonomics).
