+++ src/network_types/session.rs
@@ -1,0 +1,27 @@
+//! Types describing HOPR Sessions and their targets.
+
+use super::types::SealedHost;
+
+/// Identity of a HOPR Session.
+pub type SessionId = crate::internal::protocol::HoprPseudonym;
+
+/// Identifies a node-local service target.
+///
+/// These are specialized [`SessionTargets`](SessionTarget::ExitNode)
+/// that are local to the Exit node and have different purposes, such as Cover Traffic.
+///
+/// These targets cannot be [sealed](SealedHost) from the Entry node.
+pub type ServiceId = u32;
+
+/// Defines what should happen with the data at the recipient where the
+/// data from the established session are supposed to be forwarded to some `target`.
+#[derive(Debug, Clone, PartialEq, Eq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub enum SessionTarget {
+    /// Target is running over UDP with the given IP address and port.
+    UdpStream(SealedHost),
+    /// Target is running over TCP with the given address and port.
+    TcpStream(SealedHost),
+    /// Target is a service directly at the exit node with the given service ID.
+    ExitNode(ServiceId),
+}
